requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
beautifulsoup4==4.13.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crcmod==1.7
cryptography==44.0.0
datasets==3.2.0
decorator==5.1.1
deepface==0.0.93
dill==0.3.8
editdistance==0.8.1
ffmpeg-python==0.2.0
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
frozenlist==1.5.0
fsspec==2024.9.0
funasr==1.2.2
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.12.1
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.5
jmespath==0.10.0
joblib==1.4.2
kaldiio==2.18.0
keras==3.8.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
modelscope==1.22.2
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
mtcnn==1.0.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pip-licenses==5.0.0
platformdirs==4.3.6
pooch==1.8.2
prettytable==3.14.0
proglog==0.1.10
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pycryptodome==3.21.0
pydub==0.25.1
Pygments==2.19.1
PyMySQL==1.1.1
pynndescent==0.5.13
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-wpe==0.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
simplejson==3.19.3
six==1.17.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.5.1
torch-complex==0.4.4
torchaudio==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
